+++ client/update_test.go
@@ -19,6 +19,8 @@ import (
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
+
+	"perun.network/go-perun/channel"
 )
 
 func TestUpdateResponder_Accept_NilArgs(t *testing.T) {
@@ -36,7 +38,7 @@ func TestUpdateResponder_Reject_NilArgs(t *testing.T) {
 // Channel.Update() is defined in `client/update.go` so its test can be found
 // here as well.
 func TestChannel_Update_NilArgs(t *testing.T) {
-	err := new(Channel).Update(nil, *new(ChannelUpdate))
+	err := new(Channel).Update(nil, new(channel.State))
 	require.Error(t, err)
 	assert.Contains(t, err.Error(), "context")
 }